requirements.txt
flask==2.2.3
redis
psycopg2-binary
scrapy
selenium